+++ lib/rules/component-name-in-template-casing.js
@@ -124,7 +124,7 @@ module.exports = {
         (!utils.isHtmlElementNode(node) && !utils.isSvgElementNode(node)) ||
         utils.isHtmlWellKnownElementName(node.rawName) ||
         utils.isSvgWellKnownElementName(node.rawName) ||
-        utils.isVueTemplateSyntaxName(node.rawName)
+        utils.isVueBuiltInElementName(node.rawName)
       ) {
         return false
       }
