+++ lib/utils/index.js
@@ -54,7 +54,7 @@ const VUE2_BUILTIN_COMPONENT_NAMES = new Set(
 const VUE3_BUILTIN_COMPONENT_NAMES = new Set(
   require('./vue3-builtin-components')
 )
-const VUE_TEMPLATE_SYNTAX_NAMES = new Set(require('./vue-template-syntax'))
+const VUE_BUILTIN_ELEMENT_NAMES = new Set(require('./vue-builtin-elements'))
 const path = require('path')
 const vueEslintParser = require('vue-eslint-parser')
 const { traverseNodes, getFallbackKeys, NS } = vueEslintParser.AST
@@ -869,12 +869,12 @@ module.exports = {
   },
 
   /**
-   * Check whether the given name is Vue builtin template syntax name or not.
+   * Check whether the given name is Vue builtin element name or not.
    * @param {string} name The name to check.
-   * @returns {boolean} `true` if the name is a builtin Vue template syntax name
+   * @returns {boolean} `true` if the name is a builtin Vue element name
    */
-  isVueTemplateSyntaxName(name) {
-    return VUE_TEMPLATE_SYNTAX_NAMES.has(name.toLowerCase())
+  isVueBuiltInElementName(name) {
+    return VUE_BUILTIN_ELEMENT_NAMES.has(name.toLowerCase())
   },
 
   /**
